requirements.txt
Django==5.2.11
djangorestframework==3.16.1
numpy==2.2.4
pandas==2.2.3
PyMuPDF==1.25.5
python-docx==1.1.2
sentence-transformers==3.4.1
torch==2.6.0
transformers==4.49.0
scikit-learn==1.6.1
scipy==1.15.2
huggingface-hub==0.29.2
tokenizers==0.21.1
safetensors==0.5.3
tqdm==4.67.1
requests==2.32.3
